requirements.txt

flask==2.1.1
Flask-Testing==0.8.1
psycopg2-binary==2.9.3
gunicorn==20.1.0
python-decouple==3.6
python-dotenv==0.20.0
pytest==7.1.1
redis==4.2.2
Flask-Caching==1.10.1